pyproject.toml
[tool.poetry]
name = "spaceship"
version = "0.1.0"
description = "Prediction which passengers are transported to an alternate dimension"
authors = ["Alina Kostromina <amkostromina@edu.hse.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.26.3"
matplotlib = "^3.8.2"
torch = "^2.1.2"
scikit-learn = "^1.3.2"
pandas = "^2.1.4"
skops = "^0.9.0"
pytorch-lightning = "2.1.2"
skl2onnx = "^1.16.0"
onnxruntime = "^1.16.3"
lightning = "2.1.2"
safetensors = "^0.4.1"
mlflow = "^2.9.2"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
dvc = "^3.38.1"
dvc-gdrive = "^3.0.1"

[tool.black]
line-length = 90
target-version = ["py39"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
